resolves issues WW-279 and WW-280
It appeared as if the code to handle the Velocity portion of the FormTag and the TokenTag were never implemented.  The FormTag was no problem to implement, however, the TokenTag because of its usage of pageContext attributes was a little messier.  Ultimately, I ended up not using the pageContext attributes.  (Need to talk to Jason to see what the impact of this is).