Committing Gabe's code to make the XWorkList / XWorkMap stuff work without having to put XWork specific collection types into the Action.
Also committing a change to make the type converter re-use the existing collection in a property instead of replacing it with a new one. This is especially needed when the collection instance has extra context information, such as collections set on domain objects by Hibernate. This way the Collection will remain the same and just the objects inside it will be changed.